Updated script manager with ScriptExecutionConfig

+++ acuity-control-client/src/main/java/com/acuity/control/client/scripts/ScriptManager.java
@@ -18,9 +18,9 @@ import java.util.stream.Collectors;
 public class ScriptManager {
 
     private final Object lock = new Object();
-    private Map<String, Pair<ScriptRunConfig, Object>> scriptInstances = new HashMap<>();
+    private Map<ScriptExecutionConfig, Object> scriptInstances = new HashMap<>();
     private ScriptQueue scriptQueue = new ScriptQueue();
-    private ScriptRunConfig currentRunConfig;
+    private Pair<ScriptExecutionConfig, Object>  currentScriptExecution;
     private BotControl controller;
 
     public ScriptManager(BotControl botControl) {
@@ -29,17 +29,20 @@ public class ScriptManager {
 
     public void onLoop(){
         if (scriptQueue.getConditionalScriptMap().size() == 0){
-            if (currentRunConfig == null){
-                currentRunConfig = null;
+            if (currentScriptExecution != null){
+                currentScriptExecution = null;
                 controller.updateCurrentScriptRunConfig(null);
             }
         }
         else {
-            for (ScriptExecutionConfig pair : scriptQueue.getConditionalScriptMap()) {
-                if (ScriptConditionEvaluator.evaluate(pair.getScriptRunCondition())){
-                    if (!isCurrentScriptRunConfig(pair.getScriptRunConfig())){
-                        currentRunConfig = pair.getScriptRunConfig();
-                        controller.updateCurrentScriptRunConfig(pair.getScriptRunConfig());
+            for (ScriptExecutionConfig executionConfig : scriptQueue.getConditionalScriptMap()) {
+                if (ScriptConditionEvaluator.evaluate(executionConfig.getScriptRunCondition())){
+                    if (!isCurrentScriptExecutionConfig(executionConfig)){
+                        Object scriptInstance = getScriptInstanceOf(executionConfig);
+                        if (scriptInstance != null){
+                            currentScriptExecution = new Pair<>(executionConfig, scriptInstance);
+                            controller.updateCurrentScriptRunConfig(executionConfig.getScriptRunConfig());
+                        }
                     }
                     return;
                 }
@@ -47,8 +50,9 @@ public class ScriptManager {
         }
     }
 
-    private boolean isCurrentScriptRunConfig(ScriptRunConfig config){
-        Integer currentConfigHashCode = currentRunConfig != null ? currentRunConfig.hashCode() : null;
+    private boolean isCurrentScriptExecutionConfig(ScriptExecutionConfig config){
+        ScriptExecutionConfig currentExecutionConfig = currentScriptExecution != null ? currentScriptExecution.getKey() : null;
+        Integer currentConfigHashCode = currentExecutionConfig != null ? currentExecutionConfig.hashCode() : null;
         Integer otherConfigHashCode = config != null ? config.hashCode() : null;
         return Objects.equals(currentConfigHashCode, otherConfigHashCode);
     }
@@ -57,37 +61,41 @@ public class ScriptManager {
         if (scriptQueue.hashCode() != botClientConfig.getScriptQueue().hashCode()) {
             this.scriptQueue = botClientConfig.getScriptQueue();
             synchronized (lock){
-                List<String> ids = scriptQueue.getConditionalScriptMap().stream().map(pair -> pair.getScriptRunConfig().getRunConfigID()).collect(Collectors.toList());
-                List<String> toRemove = scriptInstances.values().stream().filter(pair -> !ids.contains(pair.getKey().getRunConfigID())).map(pair -> pair.getKey().getRunConfigID()).collect(Collectors.toList());
+                List<ScriptExecutionConfig> conditionalScriptMap = scriptQueue.getConditionalScriptMap();
+                List<ScriptExecutionConfig> toRemove = scriptInstances.keySet().stream().filter(executionConfig -> !conditionalScriptMap.contains(executionConfig)).collect(Collectors.toList());
                 toRemove.forEach(s -> scriptInstances.remove(s));
             }
         }
     }
 
-    public Pair<ScriptRunConfig, Object> getScriptInstance(){
-        ScriptRunConfig currentRunConfig = this.currentRunConfig;
-        if (currentRunConfig != null) {
-            if (!scriptInstances.containsKey(currentRunConfig.getRunConfigID())) {
+    private Object getScriptInstanceOf(ScriptExecutionConfig executionConfig){
+        if (executionConfig != null) {
+            if (!scriptInstances.containsKey(executionConfig)) {
                 synchronized (lock){
-                    Object instanceOfScript = controller.createInstanceOfScript(currentRunConfig);
+                    Object instanceOfScript = controller.createInstanceOfScript(executionConfig.getScriptRunConfig());
                     if (instanceOfScript != null) {
-                        Pair<ScriptRunConfig, Object> pair = new Pair<>(currentRunConfig, instanceOfScript);
-                        scriptInstances.put(currentRunConfig.getRunConfigID(), pair);
-                        return pair;
+                        scriptInstances.put(executionConfig, instanceOfScript);
+                        return instanceOfScript;
                     }
                 }
             }
-            return scriptInstances.get(currentRunConfig.getRunConfigID());
+            return scriptInstances.get(executionConfig);
         }
         return null;
     }
 
-    public void onScriptEnded(Pair<ScriptRunConfig, Object> closedScript) {
+    public Pair<ScriptExecutionConfig, Object> getScriptInstance(){
+        return currentScriptExecution;
+    }
+
+    public void onScriptEnded(Pair<ScriptExecutionConfig, Object> closedScript) {
         synchronized (lock){
-            scriptQueue.getConditionalScriptMap().removeIf(pair -> pair.getScriptRunConfig().getRunConfigID().equals(closedScript.getKey().getRunConfigID()));
-            controller.updateScriptQueue(scriptQueue).waitForResponse(15, TimeUnit.SECONDS);
-            scriptInstances.remove(closedScript.getKey().getRunConfigID());
-            controller.destroyInstanceOfScript(closedScript.getValue());
+            if (closedScript.getKey().isRemoveOnEnd()){
+                scriptQueue.getConditionalScriptMap().removeIf(pair -> pair.getScriptRunConfig().getRunConfigID().equals(closedScript.getKey().getScriptRunConfig().getRunConfigID()));
+                controller.updateScriptQueue(scriptQueue).waitForResponse(15, TimeUnit.SECONDS);
+                scriptInstances.remove(closedScript.getKey());
+                controller.destroyInstanceOfScript(closedScript.getValue());
+            }
         }
     }
 }
